Multi subscription network hub support
NOTE: there was a similar patch made upstream here:
https://github.com/aztfmod/terraform-azurerm-caf/pull/718/files
We would need to align our approaches before submitting this.

TL;DR
-----
To create a network hub / spoke across two subscriptions:
* Follow the existing examples but add 'provision_in_hub=true' to the resources on the hub end
* Set 'hub_subscription_id=xxx' in your launchpad global_settings

Summary
-------
This module supports the creation of a network hub (vnet, firewall,
peering, etc.), but only in a single subscription. This is fine for demo
purposes but most production architectures will have the hub and spokes
in separate subscriptions.

While one can implement this architecture by getting creative with
multiple separate landing zones and rover invocations against the
separate subscriptions, that approach is tedious and not documented.

Implementation
--------------
This commit makes use of terraform provider aliases to optionally allow
resources to be created in a separate subscription. So far the following
resources have been patched with this approach:

* firewalls
* vnets
* resource groups
* ip groups
* public ips
* vnet peering
* firewall rules

This is enough to allow the provisioning of a shared network hub in one
subscription and and peer spokes to it in another.

The same approach could be extended for other resources that may need to
be shared.

How it works
------------
As this commit only implements the above resources, here is how to add
this approach for any given resource. Example is shown for resource groups:

1. In resource_groups.tf, copy the module definition, appending "_hub"
   to the end name. In the hub block, add the provider override:

  module "resource_groups" {
    ...
  }

  module "resource_groups_hub" {
    ...

    providers = {
      azurerm = azurerm.network_hub
    }
  }

1.1 Note, if the resources are not in a submodule (e.g. as was the case
  for vnet_peerings in this commit), you will need to move them into one
  in order to make use of the new provider alias.

1.2 For some submodules (e.g. firewall in this commit), they are missing
  a definition for azurerm in their required_providers block, so this
  needs to be added for the alias to work.

2. Add/update the for_each condition in each module block to operate on
   the correct resources. Config for resource groups containing the
   attribute 'provision_in_hub = true' will be handled by the
   "resource_groups_hub" module call. Those set to false or missing this
   attribute entirely will be handled by the existing module:

  module "resource_groups" {
    for_each = {
      for key, value in try(var.resource_groups, {}) : key => value
      if try(value.provision_in_hub, false) == false
    }
    ...
  }

  module "resource_groups_hub" {
    for_each = {
      for key, value in try(var.resource_groups, {}) : key => value
      try(value.provision_in_hub, null) == true
    }
    ...
  }

3. Create a local var to combine the existing module and new module
   together:

   locals {
     resource_groups = merge(module.resource_groups, module.resource_group_reused, module.resource_groups_hub)
   }

4. Update outputs and references to module.resource_groups to
   local.resource_groups.

The main downside of this approach is that it requires the module block
to be duplicated, one for subscription a, and one for subscription b.
Right now this still seems to be the best choice:

* Provider aliases cannot use variables, so applying it per resource for
  a module would require duplicating every resource within that module.

* Applying the override at the caf_launchpad/caf_solution level is too
  early, doing this would mean every resource in a given lz would target
  the same subscription.